requirements.txt
asttokens==2.0.5
backcall==0.2.0
Bottleneck==1.3.4
cachetools==5.0.0
certifi==2020.6.20
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.3
executing==0.8.3
fonttools==4.25.0
google-api-core==2.7.1
google-auth==2.6.3
google-cloud-bigquery==3.0.1
google-cloud-bigquery-storage==2.13.1
google-cloud-core==2.2.3
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
idna==3.3
ipykernel==6.9.1
ipython==8.2.0
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.3.1
matplotlib==3.5.1
matplotlib-inline==0.1.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
nest-asyncio==1.5.1
numexpr==2.8.1
numpy==1.21.5
packaging==21.3
pandas==1.4.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.1
pip==21.2.4
prompt-toolkit==3.0.20
proto-plus==1.20.3
protobuf==3.20.0
pure-eval==0.2.2
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.11.2
pyparsing==3.0.4
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.3
pywin32==302
pyzmq==22.3.0
requests==2.27.1
rsa==4.8
scipy==1.7.3
seaborn==0.11.2
setuptools==58.0.4
sip==4.19.13
six==1.16.0
stack-data==0.2.0
tornado==6.1
traitlets==5.1.1
urllib3==1.26.9
wcwidth==0.2.5
wheel==0.37.1
wincertstore==0.2
